+++ jgclark.Dashboard/src/dashboardHelpers.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Dashboard plugin helper functions
-// Last updated 9.7.2023 for v0.5.x by @jgclark
+// Last updated 17.7.2023 for v0.5.x by @jgclark
 //-----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
@@ -9,9 +9,11 @@ import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
 import { RE_EVENT_ID } from '@helpers/calendar'
 import { getDateStringFromCalendarFilename, toLocaleTime } from '@helpers/dateTime'
 import { prependTodoToCalendarNote } from '@helpers/NPParagraph'
+import { isTermInURL } from '@helpers/paragraph'
 import {
   RE_EVENT_LINK,
   RE_MARKDOWN_LINKS_CAPTURE_G,
+  RE_NOTELINK_G,
   RE_SCHEDULED_DATES_G,
   RE_SYNC_MARKER
 } from '@helpers/regex'
@@ -90,11 +92,11 @@ export async function getSettings(): Promise<any> {
     // Set local pref Dashboard-filterPriorityItems to default false
     // if it doesn't exist already
     const savedValue = DataStore.preference('Dashboard-filterPriorityItems')
-    logDebug(pluginJson, `filter? savedValue: ${String(savedValue)}`)
+    // logDebug(pluginJson, `filter? savedValue: ${String(savedValue)}`)
     if (!savedValue) {
       DataStore.setPreference('Dashboard-filterPriorityItems', false)
     }
-    logDebug(pluginJson, `filter? -> ${String(DataStore.preference('Dashboard-filterPriorityItems'))}`)
+    // logDebug(pluginJson, `filter? -> ${String(DataStore.preference('Dashboard-filterPriorityItems'))}`)
     return config
   } catch (err) {
     logError(pluginJson, `${err.name}: ${err.message}`)
@@ -113,6 +115,7 @@ export async function getSettings(): Promise<any> {
  * - turns #hashtags and @mentions the colour that the theme displays them
  * - turns >date markers the colour that the theme displays them
  * Further, if noteTitle is supplied, then either 'append' it as a active NP note title, or make it the active NP note link for 'all' the string.
+ * Note: the actual note link is added following load by adding click handler to all items with class "sectionItemContent" (which already have a basic <a>...</a> wrapper).
  * @author @jgclark
  * @param {SectionItem} thisItem
  * @param {string?} noteTitle
@@ -121,15 +124,15 @@ export async function getSettings(): Promise<any> {
  */
 export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, noteTitle: string = "", noteLinkStyle: string = "all"): string {
   try {
-    // logDebug(`makeParaCTLLNPDIH`, `for '${thisItem.ID}' / noteTitle '${noteTitle}' filename '${thisItem.filename}'`)
+    logDebug(`makeParaContent...`, `for '${thisItem.ID}' / noteTitle '${noteTitle}' / filename '${thisItem.filename}'`)
     // Start with the content of the item
     let output = thisItem.content
 
     if (noteTitle === '(error)') {
-      logError('makeParaCTLLNDIH', `starting with noteTitle '(error)' for '${thisItem.content}'`)
+      logError('makeParaContent...', `starting with noteTitle '(error)' for '${thisItem.content}'`)
     }
 
-    // TODO: Simplify NP event links
+    // Simplify NP event links
     // of the form `![📅](2023-01-13 18:00:::F9766457-9C4E-49C8-BC45-D8D821280889:::NA:::Contact X about Y:::#63DA38)`
     let captures = output.match(RE_EVENT_LINK)
     if (captures) {
@@ -143,6 +146,16 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
       }
     }
 
+    // Simplify embedded images of the form ![image](...) by replacing with an icon.
+    // (This also helps remove false positives for ! priority indicator)
+    captures = output.match(/!\[image\]\([^\)]+\)/g)
+    if (captures) {
+      clo(captures, 'results from embedded image match:')
+      for (const capture of captures) {
+        output = output.replace(capture, `<i class="fa-regular fa-image"></i> `)
+      }
+    }
+
     // Display markdown links of the form [title](URI) as HTML links
     output = changeMarkdownLinksToHTMLLink(output)
 
@@ -153,13 +166,15 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
     // Note: need to make only one capture group, and use 'g'lobal flag
     // captures = output.match(/(\s|^|\"|\'|\(|\[|\{)(?!#[\d[:punct:]]+(\s|$))(#([^[:punct:]\s]|[\-_\/])+?\(.*?\)|#([^[:punct:]\s]|[\-_\/])+)/) // regex from @EduardMe's file
     // captures = output.match(/(\s|^|\"|\'|\(|\[|\{)(?!#[\d\'\"]+(\s|$))(#([^\'\"\s]|[\-_\/])+?\(.*?\)|#([^\'\"\s]|[\-_\/])+)/) // regex from @EduardMe's file without :punct:
-    captures = output.match(/(?:\s|^|\"|\(|\)|\')(#[A-Za-z]\w*)/g) // my too-simple version [needs to allow starting digit, just not all digits]
+    captures = output.match(/\B(?:#|＃)((?![\p{N}_]+(?:$|\b|\s))(?:[\p{L}\p{M}\p{N}_]{1,60}))/ug) // copes with Unicode characters, with help from https://stackoverflow.com/a/74926188/3238281
     // captures = output.match(HASHTAG_STR_FOR_JS) // TODO: from EM
     if (captures) {
       // clo(captures, 'results from hashtag matches:')
       for (const capture of captures) {
-        const match = capture.slice()
-        output = output.replace(match, `<span class="hashtag">${match}</span>`)
+        logDebug('makeParaContet...', `capture: ${capture}`)
+        if (!isTermInNotelinkOrURI(output, capture)) {
+          output = output.replace(capture, `<span class="hashtag">${capture}</span>`)
+        }
       }
     }
 
@@ -167,12 +182,12 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
     // Note: need to make only one capture group, and use 'g'lobal flag
     // captures = output.match(/(\s|^|\"|\'|\(|\[|\{)(?!@[\d[:punct:]]+(\s|$))(@([^[:punct:]\s]|[\-_\/])+?\(.*?\)|@([^[:punct:]\s]|[\-_\/])+)/) // regex from @EduardMe's file
     // captures = output.match(/(\s|^|\"|\'|\(|\[|\{)(?!@[\d\`\"]+(\s|$))(@([^\`\"\s]|[\-_\/])+?\(.*?\)|@([^\`\"\s]|[\-_\/])+)/) // regex from @EduardMe's file, without [:punct:]
-    captures = output.match(/(?:\s|^|\"|\(|\)\')(@[A-Za-z][\w\d\.\-\(\)]*)/g) // my too-simple version
+    captures = output.match(/\B@((?![\p{N}_]+(?:$|\b|\s))(?:[\p{L}\p{M}\p{N}_]{1,60}))/ug) // copes with Unicode characters, with help from https://stackoverflow.com/a/74926188/3238281
     // captures = output.match(NP_RE_attag_G) // TODO: from EM
     if (captures) {
-      // clo(captures, 'results from mention matches:')
+      clo(captures, 'results from mention matches:')
       for (const capture of captures) {
-        const match = capture.slice()
+        const match = capture//[2] // part from @
         output = output.replace(match, `<span class="attag">${match}</span>`)
       }
     }
@@ -196,11 +211,36 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
     // Strip all `<YYYY-MM-DD` dates
     output = stripBackwardsDateRefsFromString(output)
 
-    // TODO: add basic ***bolditalic*** styling
+    // add basic ***bolditalic*** styling
+    const RE_BOLD_ITALIC_PHRASE = new RegExp(/\*\*\*\b(.*?)\b\*\*\*/, "g")
+    captures = output.matchAll(RE_BOLD_ITALIC_PHRASE)
+    if (captures) {
+      for (const capture of captures) {
+        // logDebug('makeParaContet...', `- making bold-italic with [${String(capture)}]`)
+        output = output.replace(capture[0], `<b><i>${capture[1]}</i></b>`)
+      }
+    }
 
-    // TODO: add basic **bold** styling
+    // add basic **bold** or __bold__ styling
+    const RE_BOLD_PHRASE = new RegExp(/([_\*]{2})([^_*]+?)\1/, "g")
+    captures = output.matchAll(RE_BOLD_PHRASE)
+    if (captures) {
+      for (const capture of captures) {
+        // logDebug('makeParaContet...', `- making bold with [${String(capture)}]`)
+        output = output.replace(capture[0], `<b>${capture[2]}</b>`)
+      }
+    }
 
-    // TODO: add basic _italic_ styling
+    // add basic *italic* or _italic_ styling
+    // Note: simplified regex needs to come after bold above
+    const RE_ITALIC_PHRASE = new RegExp(/([_\*])([^*]+?)\1/, "g")
+    captures = output.matchAll(RE_ITALIC_PHRASE)
+    if (captures) {
+      for (const capture of captures) {
+        // logDebug('makeParaContet...', `- making italic with [${String(capture)}]`)
+        output = output.replace(capture[0], `<i>${capture[2]}</i>`)
+      }
+    }
 
     // Add suitable colouring to remaining >date items
     // Note: This is my attempt at finding all scheduled date links
@@ -218,16 +258,17 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
     captures = output.match(/\[\[(.*?)\]\]/)
     if (captures) {
       // clo(captures, 'results from [[notelinks]] match:')
-      for (const capturedTitle of captures) {
-        // logDebug('makeParaCTLLNDIH', `- making notelink with ${thisItem.filename}, ${capture}`)
+      for (let capturedTitle of captures) {
+        logDebug('makeParaContet...', `- making notelink with ${thisItem.filename}, ${capturedTitle}`)
+        // Replace [[notelinks]] with HTML equivalent, aware that this will interrupt the <a>...</a> that will come around the whole string, and so it needs to make <a>...</a> regions for the rest of the string before and after the capture.
         const noteTitleWithOpenAction = makeNoteTitleWithOpenActionFromTitle(capturedTitle)
-        output = output.replace('[[' + capturedTitle + ']]', noteTitleWithOpenAction)
+        output = output.replace('[[' + capturedTitle + ']]', '</a>' + noteTitleWithOpenAction + '<a>')
       }
     }
 
     // Now include an active link to the note, if 'noteTitle' is given
     if (noteTitle) {
-      // logDebug('makeParaContet...', `- before '${noteLinkStyle}' for <${noteTitle}> ${output}`)
+      // logDebug('makeParaContet...', `- before '${noteLinkStyle}' for ${noteTitle} / {${output}}`)
       switch (noteLinkStyle) {
         case 'append': {
           output = addNoteOpenLinkToString(thisItem, output) + ' ' + makeNoteTitleWithOpenActionFromFilename(thisItem, noteTitle)
@@ -238,12 +279,11 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
           break
         }
       }
-      // logDebug('makeParaContet...', `- after:  ${output}`)
+      logDebug('makeParaContet...', `- after: '${noteLinkStyle}' for ${noteTitle} / {${output}}`)
     }
 
-    // If there's a ! !! or !!! add priorityN styling
+    // If there's a ! !! or !!! add priorityN styling around the whole string. Note: this wrapping needs to go last.
     // (Simpler regex possible as the count comes later)
-    // Note: this wrapping needs to go last
     const numExclamations = getTaskPriority(output)
     switch (numExclamations) {
       case 1: {
@@ -264,17 +304,23 @@ export function makeParaContentToLookLikeNPDisplayInHTML(thisItem: SectionItem, 
       }
     }
 
+    // logDebug('makeParaContet...', `\n-> ${output}`)
     return output
   }
   catch (error) {
-    logError('makeParaContentToLookLikeNPDisplayInHTML', `${error.name} ${error.message}`)
+    logError('makeParaContentToLookLikeNPDisplayInHTML', error.message)
     return ''
   }
 }
 
+/**
+ * Get number of consecutive '!' in 'content' that aren't at the start/end/middle of a word, or preceding a '['
+ * @param {string} content
+ * @returns number of !
+ */
 export function getTaskPriority(content: string): number {
   let numExclamations = 0
-  if (content.match(/\B\!+\B/)) {
+  if (content.match(/\B\!+\B(?!\[)/)) {
     // $FlowFixMe(incompatible-use)
     numExclamations = content.match(/\B\!+\B/)[0].length
   }
@@ -282,17 +328,13 @@ export function getTaskPriority(content: string): number {
 }
 
 /**
- * v2: Make an HTML link showing displayStr, but with href onClick event to show open the 'item' in editor and select the given line content
- * v1: Used to use x-callback method to open a Note (via its 'noteTitle' parameter)
+ * Make an HTML link showing displayStr, but with href onClick event to show open the 'item' in editor and select the given line content
  * @param {SectionItem} item's details, with raw
  * @param {string} displayStr
  * @returns {string} transformed output
  */
 export function addNoteOpenLinkToString(item: SectionItem | SectionDetails, displayStr: string): string {
   try {
-    // Method 1: x-callback
-    // return `<a href="noteplan://x-callback-url/openNote?noteTitle=${titleEncoded}">${displayStr}</a>`
-
     // Method 2: pass request back to plugin
     const filenameEncoded = encodeURIComponent(item.filename)
 
@@ -313,8 +355,8 @@ export function addNoteOpenLinkToString(item: SectionItem | SectionDetails, disp
 }
 
 /**
- * v2: Wrap string with href onClick event to show note in editor
- * v1: Used to use x-callback method to open a Note (via ite 'noteTitle')
+ * Wrap string with href onClick event to show note in editor,
+ * using item.filename param.
  * @param {SectionItem} item's details
  * @param {string} noteTitle
  * @returns {string} output
@@ -322,14 +364,7 @@ export function addNoteOpenLinkToString(item: SectionItem | SectionDetails, disp
 export function makeNoteTitleWithOpenActionFromFilename(item: SectionItem, noteTitle: string): string {
   try {
     logDebug('makeNoteTitleWithOpenActionFromFilename', `- making notelink with ${item.filename}, ${noteTitle}`)
-    // Method 1: x-callback
-    // const titleEncoded = encodeURIComponent(noteTitle)
-    // return `<span class="noteTitle sectionItem"><i class="fa-regular fa-file-lines"></i> <a href="noteplan://x-callback-url/openNote?noteTitle=${titleEncoded}">${noteTitle}</a></span>`
-
-    // Method 2: pass request back to plugin
-    // Note: not passing rawContent (param 4) as its not needed
-    // return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem('${item.ID}','showNoteInEditorFromFilename','${item.filename}','')"><i class="fa-regular fa-file-lines"></i> ${noteTitle}</a>`
-    // Method 3: pass as a single object
+    // Pass request back to plugin, as a single object
     return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem({itemID: '${item.ID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(item.filename)}', encodedContent: ''})"><i class="fa-regular fa-file-lines"></i> ${noteTitle}</a>`
   }
   catch (error) {
@@ -339,7 +374,8 @@ export function makeNoteTitleWithOpenActionFromFilename(item: SectionItem, noteT
 }
 
 /**
- * Wrap string with href onClick event to show note in editor
+ * Wrap string with href onClick event to show note in editor,
+ * using noteTitle param.
  * Note: based only on 'noteTitle', not a filename
  * @param {string} noteTitle
  * @returns {string} output
@@ -349,11 +385,60 @@ export function makeNoteTitleWithOpenActionFromTitle(noteTitle: string): string 
     logDebug('makeNoteTitleWithOpenActionFromTitle', `- making notelink from ${noteTitle}`)
     // Pass request back to plugin
     // Note: not passing rawContent (param 4) as its not needed
-    // return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem('fake','showNoteInEditorFromTitle','${encodeURIComponent(noteTitle)}','')"><i class="fa-regular fa-file-lines"></i> ${noteTitle}</a>`
     return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem({itemID:'fake', type:'showNoteInEditorFromTitle', encodedFilename:'${encodeURIComponent(noteTitle)}', encodedContent:''})"><i class="fa-regular fa-file-lines"></i> ${noteTitle}</a>`
   }
   catch (error) {
     logError('makeNoteTitleWithOpenActionFromTitle', `${error.message} for input '${noteTitle}'`)
     return '(error)'
   }
+}
+
+/**
+ * Is 'term' (typically a #tag) found in [[...]] or a URL in a string which may contain 0 or more notelinks and URLs?
+ * @param {string} input
+ * @param {string} term
+ * @returns {boolean} true if found
+ */
+export function isTermInNotelinkOrURI(input: string, term: sting): boolean {
+  if (term === '') {
+    logDebug(`isTermInNotelinkOrURI`, `empty search term`)
+    return false
+  }
+  if (input === '') {
+    logDebug(`isTermInNotelinkOrURI`, `empty input string to search`)
+    return false
+  }
+  // Where is the term in the input?
+  const index = input.indexOf(term)
+  if (index < 0) {
+    logDebug(`isTermInNotelinkOrURI`, `term ${term} not found in'${input}'`)
+    return false
+  }
+  // Find any [[...]] ranges
+  let matches = input.matchAll(RE_NOTELINK_G)
+  if (matches) {
+    for (const match of matches) {
+      clo(match)
+      let rangeStart = match.index
+      let rangeEnd = match.index + match[0].length
+      logDebug(`isTermInNotelinkOrURI`, `[[...]] range: ${String(rangeStart)}-${String(rangeEnd)}`)
+      if (index >= rangeStart && index <= rangeEnd) {
+        return true
+      }
+    }
+  }
+  // Check for URL ranges. Following isn't perfect, but close enough for URLs on their own or in a [markdown](link).
+  return isTermInURL(term, input)
+}
+
+/** tests for above function */
+function testTermInNotelinkOrURI() {
+  logDebug('test1 -> false', String(isTermInNotelinkOrURI('[[link with#tag]] but empty search term', '')))
+  logDebug('test2 -> true', String(isTermInNotelinkOrURI('[[link with#tag]]', '#tag')))
+  logDebug('test3 -> false', String(isTermInNotelinkOrURI('[[link without that tag]]', '#tag')))
+  logDebug('test4 -> false', String(isTermInNotelinkOrURI('string has #tag [[but link without]]', '#tag')))
+  logDebug('test5 -> false', String(isTermInNotelinkOrURI('string has [[but link without]] and  #tag after', '#tag')))
+  logDebug('test6 -> true', String(isTermInNotelinkOrURI('term is in URL http://bob.com/page#tag', '#tag')))
+  logDebug('test7 -> false', String(isTermInNotelinkOrURI('string has http://bob.com/page #tag', '#tag')))
+  logDebug('test8 -> false', String(isTermInNotelinkOrURI('string has #tag before not in http://bob.com/URL', '#tag')))
 }